+++ packages/kit/src/runtime/client/client.js
@@ -1,5 +1,5 @@
 import { onMount, tick } from 'svelte';
-import { make_trackable, decode_params, normalize_path } from '../../utils/url.js';
+import { make_trackable, decode_params, normalize_path, add_data_suffix } from '../../utils/url.js';
 import { find_anchor, get_base_uri, scroll_state } from './utils.js';
 import {
 	lock_fetch,
@@ -14,7 +14,6 @@ import Root from '__GENERATED__/root.svelte';
 import { nodes, server_loads, dictionary, matchers, hooks } from '__GENERATED__/client-manifest.js';
 import { HttpError, Redirect } from '../control.js';
 import { stores } from './singletons.js';
-import { DATA_SUFFIX } from '../../constants.js';
 import { unwrap_promises } from '../../utils/promises.js';
 import * as devalue from 'devalue';
 
@@ -80,7 +79,7 @@ export function create_client({ target, base, trailing_slash }) {
 	/** @type {Array<((url: URL) => boolean)>} */
 	const invalidated = [];
 
-	/** @type {{id: string, promise: Promise<import('./types').NavigationResult | undefined>} | null} */
+	/** @type {{id: string, promise: Promise<import('./types').NavigationResult>} | null} */
 	let load_cache = null;
 
 	const callbacks = {
@@ -104,6 +103,9 @@ export function create_client({ target, base, trailing_slash }) {
 	let started = false;
 	let autoscroll = true;
 	let updating = false;
+	let navigating = false;
+	let hash_navigating = false;
+
 	let force_invalidation = false;
 
 	/** @type {import('svelte').SvelteComponent} */
@@ -132,8 +134,6 @@ export function create_client({ target, base, trailing_slash }) {
 		history.scrollRestoration = 'manual';
 		scrollTo(scroll.x, scroll.y);
 	}
-
-	let hash_navigating = false;
 
 	/** @type {import('types').Page} */
 	let page;
@@ -164,13 +164,19 @@ export function create_client({ target, base, trailing_slash }) {
 
 	/**
 	 * @param {string | URL} url
-	 * @param {{ noscroll?: boolean; replaceState?: boolean; keepfocus?: boolean; state?: any }} opts
+	 * @param {{ noScroll?: boolean; replaceState?: boolean; keepFocus?: boolean; state?: any; invalidateAll?: boolean }} opts
 	 * @param {string[]} redirect_chain
 	 * @param {{}} [nav_token]
 	 */
 	async function goto(
 		url,
-		{ noscroll = false, replaceState = false, keepfocus = false, state = {} },
+		{
+			noScroll = false,
+			replaceState = false,
+			keepFocus = false,
+			state = {},
+			invalidateAll = false
+		},
 		redirect_chain,
 		nav_token
 	) {
@@ -180,15 +186,19 @@ export function create_client({ target, base, trailing_slash }) {
 
 		return navigate({
 			url,
-			scroll: noscroll ? scroll_state() : null,
-			keepfocus,
+			scroll: noScroll ? scroll_state() : null,
+			keepfocus: keepFocus,
 			redirect_chain,
 			details: {
 				state,
 				replaceState
 			},
 			nav_token,
-			accepted: () => {},
+			accepted: () => {
+				if (invalidateAll) {
+					force_invalidation = true;
+				}
+			},
 			blocked: () => {},
 			type: 'goto'
 		});
@@ -223,8 +233,12 @@ export function create_client({ target, base, trailing_slash }) {
 		if (!navigation_result) {
 			navigation_result = await server_fallback(
 				url,
-				null,
-				handle_error(new Error(`Not found: ${url.pathname}`), { url, params: {}, routeId: null }),
+				{ id: null },
+				handle_error(new Error(`Not found: ${url.pathname}`), {
+					url,
+					params: {},
+					route: { id: null }
+				}),
 				404
 			);
 		}
@@ -240,9 +254,9 @@ export function create_client({ target, base, trailing_slash }) {
 			if (redirect_chain.length > 10 || redirect_chain.includes(url.pathname)) {
 				navigation_result = await load_root_error_page({
 					status: 500,
-					error: handle_error(new Error('Redirect loop'), { url, params: {}, routeId: null }),
+					error: handle_error(new Error('Redirect loop'), { url, params: {}, route: { id: null } }),
 					url,
-					routeId: null
+					route: { id: null }
 				});
 			} else {
 				goto(
@@ -373,10 +387,11 @@ export function create_client({ target, base, trailing_slash }) {
 			from: null,
 			to: add_url_properties('to', {
 				params: current.params,
-				routeId: current.route?.id ?? null,
+				route: { id: current.route?.id ?? null },
 				url: new URL(location.href)
 			}),
-			type: 'load'
+			willUnload: false,
+			type: 'enter'
 		};
 		callbacks.after_navigate.forEach((fn) => fn(navigation));
 
@@ -455,7 +470,7 @@ export function create_client({ target, base, trailing_slash }) {
 			result.props.page = {
 				error,
 				params,
-				routeId: route && route.id,
+				route,
 				status,
 				url,
 				form,
@@ -493,12 +508,12 @@ export function create_client({ target, base, trailing_slash }) {
 	 * 	 parent: () => Promise<Record<string, any>>;
 	 *   url: URL;
 	 *   params: Record<string, string>;
-	 *   routeId: string | null;
+	 *   route: { id: string | null };
 	 * 	 server_data_node: import('./types').DataNode | null;
 	 * }} options
 	 * @returns {Promise<import('./types').BranchNode>}
 	 */
-	async function load_node({ loader, parent, url, params, routeId, server_data_node }) {
+	async function load_node({ loader, parent, url, params, route, server_data_node }) {
 		/** @type {Record<string, any> | null} */
 		let data = null;
 
@@ -507,6 +522,7 @@ export function create_client({ target, base, trailing_slash }) {
 			dependencies: new Set(),
 			params: new Set(),
 			parent: false,
+			route: false,
 			url: false
 		};
 
@@ -523,7 +539,12 @@ export function create_client({ target, base, trailing_slash }) {
 
 			/** @type {import('types').LoadEvent} */
 			const load_input = {
-				routeId,
+				route: {
+					get id() {
+						uses.route = true;
+						return route.id;
+					}
+				},
 				params: new Proxy(params, {
 					get: (target, key) => {
 						uses.params.add(/** @type {string} */ (key));
@@ -608,6 +629,12 @@ export function create_client({ target, base, trailing_slash }) {
 						);
 					},
 					enumerable: false
+				},
+				routeId: {
+					get() {
+						throw new Error('routeId has been replaced by route.id');
+					},
+					enumerable: false
 				}
 			});
 
@@ -634,17 +661,19 @@ export function create_client({ target, base, trailing_slash }) {
 	}
 
 	/**
-	 * @param {boolean} url_changed
 	 * @param {boolean} parent_changed
+	 * @param {boolean} route_changed
+	 * @param {boolean} url_changed
 	 * @param {import('types').Uses | undefined} uses
 	 * @param {Record<string, string>} params
 	 */
-	function has_changed(url_changed, parent_changed, uses, params) {
+	function has_changed(parent_changed, route_changed, url_changed, uses, params) {
 		if (force_invalidation) return true;
 
 		if (!uses) return false;
 
 		if (uses.parent && parent_changed) return true;
+		if (uses.route && route_changed) return true;
 		if (uses.url && url_changed) return true;
 
 		for (const param of uses.params) {
@@ -672,6 +701,7 @@ export function create_client({ target, base, trailing_slash }) {
 					dependencies: new Set(node.uses.dependencies ?? []),
 					params: new Set(node.uses.params ?? []),
 					parent: !!node.uses.parent,
+					route: !!node.uses.route,
 					url: !!node.uses.url
 				}
 			};
@@ -683,7 +713,7 @@ export function create_client({ target, base, trailing_slash }) {
 
 	/**
 	 * @param {import('./types').NavigationIntent} intent
-	 * @returns {Promise<import('./types').NavigationResult | undefined>}
+	 * @returns {Promise<import('./types').NavigationResult>}
 	 */
 	async function load_route({ id, invalidating, url, params, route }) {
 		if (load_cache?.id === id) {
@@ -704,6 +734,7 @@ export function create_client({ target, base, trailing_slash }) {
 		let server_data = null;
 
 		const url_changed = current.url ? id !== current.url.pathname + current.url.search : false;
+		const route_changed = current.route ? id !== current.route.id : false;
 
 		const invalid_server_nodes = loaders.reduce((acc, loader, i) => {
 			const previous = current.branch[i];
@@ -711,7 +742,13 @@ export function create_client({ target, base, trailing_slash }) {
 			const invalid =
 				!!loader?.[0] &&
 				(previous?.loader !== loader[1] ||
-					has_changed(url_changed, acc.some(Boolean), previous.server?.uses, params));
+					has_changed(
+						acc.some(Boolean),
+						route_changed,
+						url_changed,
+						previous.server?.uses,
+						params
+					));
 
 			acc.push(invalid);
 			return acc;
@@ -723,9 +760,9 @@ export function create_client({ target, base, trailing_slash }) {
 			} catch (error) {
 				return load_root_error_page({
 					status: 500,
-					error: handle_error(error, { url, params, routeId: route.id }),
+					error: handle_error(error, { url, params, route: { id: route.id } }),
 					url,
-					routeId: route.id
+					route
 				});
 			}
 
@@ -750,7 +787,7 @@ export function create_client({ target, base, trailing_slash }) {
 			const valid =
 				(!server_data_node || server_data_node.type === 'skip') &&
 				loader[1] === previous?.loader &&
-				!has_changed(url_changed, parent_changed, previous.shared?.uses, params);
+				!has_changed(parent_changed, route_changed, url_changed, previous.shared?.uses, params);
 			if (valid) return previous;
 
 			parent_changed = true;
@@ -764,7 +801,7 @@ export function create_client({ target, base, trailing_slash }) {
 				loader: loader[1],
 				url,
 				params,
-				routeId: route.id,
+				route,
 				parent: async () => {
 					const data = {};
 					for (let j = 0; j < i; j += 1) {
@@ -812,7 +849,7 @@ export function create_client({ target, base, trailing_slash }) {
 						status = err.status;
 						error = err.body;
 					} else {
-						error = handle_error(err, { params, url, routeId: route.id });
+						error = handle_error(err, { params, url, route: { id: route.id } });
 					}
 
 					const error_load = await load_nearest_error_page(i, branch, errors);
@@ -828,7 +865,7 @@ export function create_client({ target, base, trailing_slash }) {
 					} else {
 						// if we get here, it's because the root `load` function failed,
 						// and we need to fall back to the server
-						return await server_fallback(url, route.id, error, status);
+						return await server_fallback(url, { id: route.id }, error, status);
 					}
 				}
 			} else {
@@ -884,11 +921,11 @@ export function create_client({ target, base, trailing_slash }) {
 	 *   status: number;
 	 *   error: App.Error;
 	 *   url: URL;
-	 *   routeId: string | null
+	 *   route: { id: string | null }
 	 * }} opts
 	 * @returns {Promise<import('./types').NavigationFinished>}
 	 */
-	async function load_root_error_page({ status, error, url, routeId }) {
+	async function load_root_error_page({ status, error, url, route }) {
 		/** @type {Record<string, string>} */
 		const params = {}; // error page does not have params
 
@@ -924,7 +961,7 @@ export function create_client({ target, base, trailing_slash }) {
 			loader: default_layout_loader,
 			url,
 			params,
-			routeId,
+			route,
 			parent: () => Promise.resolve({}),
 			server_data_node: create_data_node(server_data_node)
 		});
@@ -980,6 +1017,49 @@ export function create_client({ target, base, trailing_slash }) {
 	/**
 	 * @param {{
 	 *   url: URL;
+	 *   type: import('types').NavigationType;
+	 *   intent?: import('./types').NavigationIntent;
+	 *   delta?: number;
+	 * }} opts
+	 */
+	function before_navigate({ url, type, intent, delta }) {
+		let should_block = false;
+
+		/** @type {import('types').Navigation} */
+		const navigation = {
+			from: add_url_properties('from', {
+				params: current.params,
+				route: { id: current.route?.id ?? null },
+				url: current.url
+			}),
+			to: add_url_properties('to', {
+				params: intent?.params ?? null,
+				route: { id: intent?.route?.id ?? null },
+				url
+			}),
+			willUnload: !intent,
+			type
+		};
+
+		if (delta !== undefined) {
+			navigation.delta = delta;
+		}
+
+		const cancellable = {
+			...navigation,
+			cancel: () => {
+				should_block = true;
+			}
+		};
+
+		callbacks.before_navigate.forEach((fn) => fn(cancellable));
+
+		return should_block ? null : navigation;
+	}
+
+	/**
+	 * @param {{
+	 *   url: URL;
 	 *   scroll: { x: number, y: number } | null;
 	 *   keepfocus: boolean;
 	 *   redirect_chain: string[];
@@ -1006,39 +1086,10 @@ export function create_client({ target, base, trailing_slash }) {
 		accepted,
 		blocked
 	}) {
-		let should_block = false;
-
 		const intent = get_navigation_intent(url, false);
+		const navigation = before_navigate({ url, type, delta, intent });
 
-		/** @type {import('types').Navigation} */
-		const navigation = {
-			from: add_url_properties('from', {
-				params: current.params,
-				routeId: current.route?.id ?? null,
-				url: current.url
-			}),
-			to: add_url_properties('to', {
-				params: intent?.params ?? null,
-				routeId: intent?.route.id ?? null,
-				url
-			}),
-			type
-		};
-
-		if (delta !== undefined) {
-			navigation.delta = delta;
-		}
-
-		const cancellable = {
-			...navigation,
-			cancel: () => {
-				should_block = true;
-			}
-		};
-
-		callbacks.before_navigate.forEach((fn) => fn(cancellable));
-
-		if (should_block) {
+		if (!navigation) {
 			blocked();
 			return;
 		}
@@ -1046,6 +1097,8 @@ export function create_client({ target, base, trailing_slash }) {
 		update_scroll_positions(current_history_index);
 
 		accepted();
+
+		navigating = true;
 
 		if (started) {
 			stores.navigating.set(navigation);
@@ -1062,6 +1115,7 @@ export function create_client({ target, base, trailing_slash }) {
 			},
 			nav_token,
 			() => {
+				navigating = false;
 				callbacks.after_navigate.forEach((fn) => fn(navigation));
 				stores.navigating.set(null);
 			}
@@ -1071,12 +1125,12 @@ export function create_client({ target, base, trailing_slash }) {
 	/**
 	 * Does a full page reload if it wouldn't result in an endless loop in the SPA case
 	 * @param {URL} url
-	 * @param {string | null} routeId
+	 * @param {{ id: string | null }} route
 	 * @param {App.Error} error
 	 * @param {number} status
 	 * @returns {Promise<import('./types').NavigationFinished>}
 	 */
-	async function server_fallback(url, routeId, error, status) {
+	async function server_fallback(url, route, error, status) {
 		if (url.origin === location.origin && url.pathname === location.pathname && !hydrated) {
 			// We would reload the same page we're currently on, which isn't hydrated,
 			// which means no SSR, which means we would end up in an endless loop
@@ -1084,7 +1138,7 @@ export function create_client({ target, base, trailing_slash }) {
 				status,
 				error,
 				url,
-				routeId
+				route
 			});
 		}
 		return await native_navigation(url);
@@ -1140,7 +1194,22 @@ export function create_client({ target, base, trailing_slash }) {
 			}
 		},
 
-		goto: (href, opts = {}) => goto(href, opts, []),
+		goto: (href, opts = {}) => {
+			// TODO remove for 1.0
+			if ('keepfocus' in opts) {
+				throw new Error(
+					'`keepfocus` has been renamed to `keepFocus` (note the difference in casing)'
+				);
+			}
+
+			if ('noscroll' in opts) {
+				throw new Error(
+					'`noscroll` has been renamed to `noScroll` (note the difference in casing)'
+				);
+			}
+
+			return goto(href, opts, []);
+		},
 
 		invalidate: (resource) => {
 			if (resource === undefined) {
@@ -1212,7 +1281,7 @@ export function create_client({ target, base, trailing_slash }) {
 					post_update();
 				}
 			} else if (result.type === 'redirect') {
-				goto(result.location, {}, []);
+				goto(result.location, { invalidateAll: true }, []);
 			} else {
 				/** @type {Record<string, any>} */
 				const props = {
@@ -1235,19 +1304,24 @@ export function create_client({ target, base, trailing_slash }) {
 			addEventListener('beforeunload', (e) => {
 				let should_block = false;
 
-				/** @type {import('types').Navigation & { cancel: () => void }} */
-				const navigation = {
-					from: add_url_properties('from', {
-						params: current.params,
-						routeId: current.route?.id ?? null,
-						url: current.url
-					}),
-					to: null,
-					type: 'unload',
-					cancel: () => (should_block = true)
-				};
+				if (!navigating) {
+					// If we're navigating, beforeNavigate was already called. If we end up in here during navigation,
+					// it's due to an external or full-page-reload link, for which we don't want to call the hook again.
+					/** @type {import('types').Navigation & { cancel: () => void }} */
+					const navigation = {
+						from: add_url_properties('from', {
+							params: current.params,
+							route: { id: current.route?.id ?? null },
+							url: current.url
+						}),
+						to: null,
+						willUnload: true,
+						type: 'leave',
+						cancel: () => (should_block = true)
+					};
 
-				callbacks.before_navigate.forEach((fn) => fn(navigation));
+					callbacks.before_navigate.forEach((fn) => fn(navigation));
+				}
 
 				if (should_block) {
 					e.preventDefault();
@@ -1325,17 +1399,22 @@ export function create_client({ target, base, trailing_slash }) {
 				)
 					return;
 
-				// Ignore if tag has
-				// 1. 'download' attribute
-				// 2. 'rel' attribute includes external
-				const rel = (a.getAttribute('rel') || '').split(/\s+/);
+				if (a.hasAttribute('download')) return;
 
-				if (a.hasAttribute('download') || rel.includes('external') || options.reload) {
+				// Ignore the following but fire beforeNavigate
+				const rel = (a.getAttribute('rel') || '').split(/\s+/);
+				if (
+					rel.includes('external') ||
+					options.reload ||
+					(is_svg_a_element ? a.target.baseVal : a.target)
+				) {
+					const navigation = before_navigate({ url, type: 'link' });
+					if (!navigation) {
+						event.preventDefault();
+					}
+					navigating = true;
 					return;
 				}
-
-				// Ignore if <a> has a target
-				if (is_svg_a_element ? a.target.baseVal : a.target) return;
 
 				// Check if new url only differs by hash and use the browser default behavior in that case
 				// This will ensure the `hashchange` event is fired
@@ -1428,15 +1507,7 @@ export function create_client({ target, base, trailing_slash }) {
 			});
 		},
 
-		_hydrate: async ({
-			status,
-			error,
-			node_ids,
-			params,
-			routeId,
-			data: server_data_nodes,
-			form
-		}) => {
+		_hydrate: async ({ status, error, node_ids, params, route, data: server_data_nodes, form }) => {
 			hydrated = true;
 
 			const url = new URL(location.href);
@@ -1452,7 +1523,7 @@ export function create_client({ target, base, trailing_slash }) {
 						loader: nodes[n],
 						url,
 						params,
-						routeId,
+						route,
 						parent: async () => {
 							const data = {};
 							for (let j = 0; j < i; j += 1) {
@@ -1471,7 +1542,7 @@ export function create_client({ target, base, trailing_slash }) {
 					status,
 					error,
 					form,
-					route: routes.find((route) => route.id === routeId) ?? null
+					route: routes.find(({ id }) => id === route.id) ?? null
 				});
 			} catch (error) {
 				if (error instanceof Redirect) {
@@ -1483,9 +1554,9 @@ export function create_client({ target, base, trailing_slash }) {
 
 				result = await load_root_error_page({
 					status: error instanceof HttpError ? error.status : 500,
-					error: handle_error(error, { url, params, routeId }),
+					error: handle_error(error, { url, params, route }),
 					url,
-					routeId
+					route
 				});
 			}
 
@@ -1501,21 +1572,35 @@ export function create_client({ target, base, trailing_slash }) {
  */
 async function load_data(url, invalid) {
 	const data_url = new URL(url);
-	data_url.pathname = url.pathname.replace(/\/$/, '') + DATA_SUFFIX;
+	data_url.pathname = add_data_suffix(url.pathname);
 
 	const res = await native_fetch(data_url.href, {
 		headers: {
 			'x-sveltekit-invalidated': invalid.map((x) => (x ? '1' : '')).join(',')
 		}
 	});
-	const server_data = await res.text();
+	const data = await res.json();
 
 	if (!res.ok) {
 		// error message is a JSON-stringified string which devalue can't handle at the top level
-		throw new Error(JSON.parse(server_data));
+		throw new Error(data);
 	}
 
-	return devalue.parse(server_data);
+	// revive devalue-flattened data
+	data.nodes?.forEach((/** @type {any} */ node) => {
+		if (node?.type === 'data') {
+			node.data = devalue.unflatten(node.data);
+			node.uses = {
+				dependencies: new Set(node.uses.dependencies ?? []),
+				params: new Set(node.uses.params ?? []),
+				parent: !!node.uses.parent,
+				route: !!node.uses.route,
+				url: !!node.uses.url
+			};
+		}
+	});
+
+	return data;
 }
 
 /**
@@ -1529,7 +1614,7 @@ function handle_error(error, event) {
 	}
 	return (
 		hooks.handleError({ error, event }) ??
-		/** @type {any} */ ({ message: event.routeId != null ? 'Internal Error' : 'Not Found' })
+		/** @type {any} */ ({ message: event.route.id != null ? 'Internal Error' : 'Not Found' })
 	);
 }
 
@@ -1564,6 +1649,15 @@ function add_url_properties(type, target) {
 			enumerable: false
 		});
 	}
+
+	Object.defineProperty(target, 'routeId', {
+		get() {
+			throw new Error(
+				`The navigation shape changed - ${type}.routeId should now be ${type}.route.id`
+			);
+		},
+		enumerable: false
+	});
 
 	return target;
 }